Extract SyntaxError
